+++ backend/canisters/group_index/impl/src/lib.rs
@@ -28,6 +28,7 @@ use types::{
 use utils::canister::{CanistersRequiringUpgrade, FailedUpgradeCount};
 use utils::env::Environment;
 use utils::idempotency_checker::IdempotencyChecker;
+use utils::migrated_user_ids::MigratedUserIds;
 
 mod guards;
 mod jobs;
@@ -329,6 +330,10 @@ struct Data {
     pub rng_seed: [u8; 32],
     pub idempotency_checker: IdempotencyChecker,
     pub local_index_event_sync_queue: GroupedTimerJobQueue<LocalIndexEventBatch>,
+    // The latest ids of migrated users, as looked up from the UserIndex whenever a user's id is found to
+    // have changed
+    #[serde(default)]
+    pub migrated_user_ids: MigratedUserIds,
 }
 
 impl Data {
@@ -374,6 +379,7 @@ impl Data {
             rng_seed: [0; 32],
             idempotency_checker: IdempotencyChecker::default(),
             local_index_event_sync_queue: GroupedTimerJobQueue::new(10, false),
+            migrated_user_ids: MigratedUserIds::default(),
         }
     }
 
@@ -482,6 +488,7 @@ impl Default for Data {
             rng_seed: [0; 32],
             idempotency_checker: IdempotencyChecker::default(),
             local_index_event_sync_queue: GroupedTimerJobQueue::new(10, false),
+            migrated_user_ids: MigratedUserIds::default(),
         }
     }
 }